Add refetch for queries
